XR Interaction Toolkit Overhaul and VR Hands
Changed the project from using the newer XR Plug-in Management system. The project will be using the built-in XR system as it will be able to support multiple platforms like Oculus and OpenVR.

The project now consists of the same  test scene as before - but with different materials applied and basic  lighting generated.

There are now various subfolders within the Assets folder that contain Materials, Models, and Prefabs for both the prototype level scene and the hand-presence prefabs.

*Tested on an Oculus Rift CV1 HMD/Touch Controllers*

Known Issues:
- The hands currently have no collision in place, so they go through any object (Waiting to have a better understanding of how collision is usually applied on tracked controllers).

Fixed Bugs:
- The camera would make a sharp "snap" when turning more than 180 degrees

References:
https://docs.unity3d.com/Packages/com.unity.xr.interaction.toolkit@0.9/manual/index.html
https://docs.unity3d.com/Manual/VROverview.html
https://docs.unity3d.com/Manual/configuring-project-for-xr.html